Refactored checkin.js into checkin_functions.js and checkin.js
