+++ app/js/checkin.js
@@ -1,5 +1,11 @@
 var  dialog  = require('electron').remote;
+const checkin_functions = require('./checkin_functions.js');
+const {checkTeams, checkPrevTask, startFlow, addTask} = checkin_functions;
 
+checkPrevTask();
+
+var startFlowButton = document.getElementById("startFlowBtn")
+startFlowButton.addEventListener("click", () => startFlow())
 
 var firebaseConfig = {
     apiKey: "AIzaSyBmn_tDSlm4lLdrvSqj8Yb00KkYae8cL-Y",
@@ -12,6 +18,7 @@ var firebaseConfig = {
     measurementId: "G-VMS6JL8H4S"
 };
 // Initialize Firebase
+var firebase = require('firebase');
 firebase.initializeApp(firebaseConfig);
 var db = firebase.firestore();
 var uid = localStorage.getItem('userid')
@@ -24,117 +31,6 @@ task2.value = ""
 task3.value = ""
 var prevList = document.getElementById("prevTask")
 var todayTask = document.getElementById("todayTask")
-
-checkTeams()
-
-function checkTeams() {
-	var errorMessage = "An error occurred when trying to find your team, returning to main page."
-    db.collection("teams").where(uid, "==", true)
-        .get()
-        .then(function(querySnapshot) {
-            console.log(querySnapshot.docs)
-            if (querySnapshot.docs.length > 0) {
-                querySnapshot.forEach(function(doc) {
-                    teamName = doc.id
-                });
-            } else {
-            	dialog.showMessageBox({
-		            type: 'error',
-		            title: 'Error',
-		            message: errorMessage
-		        });
-                console.log("Team not found")
-                document.location.href = 'taskbar.html'
-            }
-        })
-        .catch(function(error) {
-        	dialog.showMessageBox({
-	            type: 'error',
-	            title: 'Error',
-	            message: errorMessage
-	        });
-            console.log("Error getting documents: ", error);
-            document.location.href = 'taskbar.html'
-        });
-}
-function checkPrevTask(){
-    db.collection("teams").where(uid, "==", true)
-        .get()
-        .then(function(querySnapshot) {
-            console.log(querySnapshot.docs)
-            if (querySnapshot.docs.length > 0) {
-                querySnapshot.forEach(function(doc) {
-                    teamName = doc.id
-                });
-            }
-            return teamName
-
-        }).then(function(teamName){
-        db.collection("teams")
-        .doc(teamName)
-        .collection(uid).doc("status").get()
-        .then(function(status){
-           statusObj = status.data()
-           ptDiv= document.getElementById("prevTask")
-           console.log(statusObj.task1)
-           if(statusObj.task1!="" || statusObj.task2!="" || statusObj.task3!=""){
-                ptDiv.style.display="block"
-           }
-           if(statusObj.task1!=""){
-               addTask(ptDiv,statusObj.task1)
-           }
-           if(statusObj.task2!=""){
-               addTask(ptDiv,statusObj.task2)
-           }
-           if(statusObj.task3!=""){
-               addTask(ptDiv,statusObj.task3)
-           }
-        })
-        .catch(function(error){
-            console.log("Error checking prev tasks",error)
-        });
-        });
-}
-checkPrevTask()
-
-var startFlowButton = document.getElementById("startFlowBtn")
-startFlowButton.addEventListener("click", () => startFlow())
-
-//startflow will always send 3 tasks value, if the user didn't not set any of them, just set the val to be ""
-function startFlow() {
-    console.log("tv1",task1.value)
-    var obj = {
-    	checkedIn: true,
-        task1: task1.value,
-        task2: task2.value,
-        task3: task3.value,
-    	taskStatus: 1
-    }
-    db.collection("teams").doc(teamName).collection(uid).doc("status").set(obj)
-        .then(function() {
-            console.log("Document written");
-            document.location.href = 'taskbar.html'
-        })
-        .catch(function(error) {
-        	dialog.showMessageBox({
-	            type: 'error',
-	            title: 'Error',
-	            message: error.message
-	        });
-            console.error("Error adding document: ", error);
-            document.location.href = 'taskbar.html'
-        });
-}
-
-function addTask(parent,text){
-    let task = `
-        <li>
-            <input style="display: inline-block;" value = "${text}">
-            <button class="bt">Add</button>
-            <button class="bt">Delete</button>
-        </li>`
-    parent.insertAdjacentHTML('beforeend', task);
-}
 
 prevList.addEventListener("click", function(event){
     let target = event.target;
@@ -197,4 +93,3 @@ todayTask.addEventListener('click',function(event){
         targetParent.firstElementChild.value=""
     }
 });
-module.exports = {checkTeams, startFlow, firebase, dialog};
