+++ package.json
@@ -38,14 +38,20 @@
   "electron": "^8.2.5",
   "electron-builder": "^22.6.0",
   "eslint": "^7.0.0",
+  "firebase-mock": "^2.3.2",
   "husky": "^4.2.5",
-  "mocha": "^7.1.2",
+  "jasmine": "^3.5.0",
+  "mocha": "^7.2.0",
+  "mock-local-storage": "^1.1.12",
   "spectron": "^10.0.1"
  },
  "dependencies": {
   "express": "^4.17.1",
   "firebase": "^7.14.4",
+  "jsdom": "^16.2.2",
   "path": "^0.12.7",
-  "ps-list": "^7.0.0"
+  "proxyquire": "^2.1.3",
+  "ps-list": "^7.0.0",
+  "sinon": "^9.0.2"
  }
 }
